+++ src/lib/constants/index.ts
@@ -1,3 +1,4 @@
+export * from './programStagePropertyLabels'
 export * from './sections'
 export * from './translatedModelConstants'
 export * from './translatedModelProperties'