+++ src/pages/programStages/List.tsx
@@ -1,0 +1,1 @@
+export { DefaultSectionList as Component } from '../DefaultSectionList'